cdk/requirements.txt
annotated-types==0.7.0
attrs==22.2.0
aws-cdk-lib==2.173.2
aws-cdk.asset-awscli-v1==2.2.215
aws-cdk.asset-kubectl-v20==2.1.3
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==38.0.1
cattrs==22.2.0
constructs==10.2.16
importlib-resources==5.12.0
jsii==1.106.0
publication==0.0.3
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
